backend/requirements.txt
flask==3.0.0
python-chess==1.9.4
gunicorn==20.1.0